+++ src/tests/excalidrawPackage.test.tsx
@@ -226,14 +226,6 @@ describe("<Excalidraw/>", () => {
   });
 
   describe("Test autofocus prop", () => {
-    it('should not focus when autofocus is "undefined"', async () => {
-      const { container } = await render(<Excalidraw />);
-
-      expect(
-        container.querySelector(".excalidraw") === document.activeElement,
-      ).toBe(false);
-    });
-
     it("should not focus when autofocus is false", async () => {
       const { container } = await render(<Excalidraw autofocus={false} />);