+++ src/packages/excalidraw/CHANGELOG.md
@@ -19,7 +19,7 @@ Please add the latest change on the top under the correct section.
 
 ### Features
 
-- Added prop `autofocus` which when enabled will focus on the Excalidraw component on page load [#3691](https://github.com/excalidraw/excalidraw/pull/3691).
+- Added prop [`autofocus`](https://github.com/excalidraw/excalidraw/blob/master/src/packages/excalidraw/README.md#autofocus) to focus the excalidraw component when enabled, defaults to false [#3691](https://github.com/excalidraw/excalidraw/pull/3691).
 - Added prop `UIOptions.canvasActions.export.renderCustomUI` to support Custom UI rendering inside export dialog [#3666](https://github.com/excalidraw/excalidraw/pull/3666).
 - Added prop `UIOptions.canvasActions.saveAsImage` to show/hide the **Save as image** button in the canvas actions. Defauls to `true` hence the **Save as Image** button is rendered [#3662](https://github.com/excalidraw/excalidraw/pull/3662).